Refactored LanguageDetector to extenda JCasAnnotator and moved it to separate package
